Create Strand Sort
Below are simple steps used in the algorithm.

Let ip[] be input list and op[] be output list.
Create an empty sublist and move first item of ip[] to it.
Traverse remaining items of ip. For every item x, check if x is greater than last inserted item to sublist. If yes, remove x from ip and add at the end of sublist. If no, ignore x (Keep it it in ip)
Merge sublist into op (output list)
Recur for remaining items in ip and current items in op.